pyproject.toml
[build-system]
requires = ["maturin>=1.0,<2.0"]
build-backend = "maturin"

[project]
name = "chess-ai"
version = "0.1.0"
description = "Chess AI tools for feature analysis and explainable gameplay"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
keywords = ["chess", "ai", "machine-learning", "explainability", "features", "engine"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "numpy>=1.20.0",
    "scikit-learn>=1.0.0",
    "tqdm>=4.60.0",
    "python-chess>=1.999",
    "flask>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
    "maturin>=1.0.0",
]

[project.scripts]
chess-ai-audit = "chess_ai.cli.audit:main"
chess-ai-play = "chess_ai.cli.explainable:main"
chess-ai = "chess_ai.cli.main:main"
chess-ai-web = "chess_ai.web.app:run_server"

[project.urls]
Homepage = "https://github.com/yourusername/chess-feature-audit"
Repository = "https://github.com/yourusername/chess-feature-audit"
Issues = "https://github.com/yourusername/chess-feature-audit/issues"

[tool.maturin]
features = ["pyo3/extension-module"]
module-name = "chess_ai.rust_utils._chess_ai_rust"
project-layout = "flat"
python-source = "src"
manifest-path = "rust/Cargo.toml"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.ruff]
line-length = 88
target-version = "py38"
extend-exclude = ["mutants"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "C90",  # mccabe complexity
    "UP",   # pyupgrade
    "S",    # flake8-bandit (security)
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific rules
    "PT",   # flake8-pytest-style
    "PIE",  # flake8-pie
    "PERF", # perflint (performance)
    "T20",  # flake8-print
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.mccabe]
max-complexity = 15

[tool.ruff.lint.per-file-ignores]
"src/chess_ai/cli/*" = ["T20"]
"src/chess_ai/explainable_engine.py" = ["T20"]
"scripts/*" = ["T20"]
"*.ipynb" = ["T20"]
"tests/*" = ["S101", "PT019"]

[tool.ruff.lint.isort]
known-first-party = ["chess_ai"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_unreachable = true
warn_redundant_casts = true
ignore_missing_imports = true
exclude = [
    "^build/",
    "^\\.venv/",
    "^venv/",
    "^\\.eggs/",
    "^dist/",
    "^mutants/",
    "^tests/",
    "^scripts/",
]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
filterwarnings = [
    "ignore::UserWarning:sklearn.*",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*",
]

[tool.mutmut]
paths_to_mutate = [
    "src/chess_ai/utils/math.py",
    "src/chess_ai/metrics/kendall.py",
    "src/chess_ai/metrics/positional.py",
    "src/chess_ai/features/utils.py",
    "src/chess_ai/engine/config.py",
    "src/chess_ai/surrogate_explainer.py",
]
also_copy = ["src/"]
backup = false
runner = "python -m pytest tests/ -p no:cov"
tests_dir = ["tests/"]